Add a separate app 'accounts' for User handling
