+++ classroom/settings.py
@@ -37,7 +37,10 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+
+    # User-added Apps
     'classnote',
+    'accounts',
 ]
 
 MIDDLEWARE = [
